Created corpse blueprints for a bunch of things that didn't previously have them.
added those corpse blueprints to the things that didn't have them.

Got the Reanimated builder working (just ironing out a kink or two with the player not *appearing* like a corpse despite being one).

created a "PastLife" part for corpses that can contain an arbitrary set of of information about the creature whose corspe it is for the purposes of reanimating it.

attempted to make a variable replacer for "weird" text