Optimised some of how the logging determines what method has been called as well as when a method should not log at all.
Got Cybernetic rejection syndrome working for the Mad Monger (or any creature with the enforcement part)

added logic for manipulating corpse tile colors post-reanimation so that both past life and corpse can influence the result.

tweaked the numbers on unending suffering. It gets worse the more times a thing is reanimated.